refactor(MCP): migrate run_agent server to static metadata pattern
- Create new metadata.ts file with static server metadata, configurable properties,
  and tool schema definitions
- Move server implementation to lib/api/actions/servers/run_agent/
- Update constants.ts to use metadata export from the new location
- Update servers/index.ts to import from new location
- Remove old index.ts file (keeping helper files in place as they're still used)

The run_agent server is unique in that its tool name is dynamically generated
based on the child agent configuration. The metadata file extracts what can
be static (server info, configurable properties schema, tool input schema)
while the tool registration remains dynamic in index.ts.
